rename some widgets
