+++ lib/groups/widgets/receipt_list_item.dart
